Наброски репозитория для работы с вкладами и кредитами. Синхронизация типов во вкладе и кредите с приходящими с сервера. Но IMHO они там странные очень.

+++ app/src/main/java/com/example/sandboxbank/App/ui/financialitem/FinancialItemInfoCard.kt
@@ -51,7 +51,7 @@ fun FinancialItemInfoCard(financialItem: FinancialItem) {
 }
 
 // Расширение для Double:
-fun Double.toFormattedBalance(): String {
+fun Long.toFormattedBalance(): String {
     val numberFormat = NumberFormat.getNumberInstance(Locale("ru", "RU")).apply {
         maximumFractionDigits = 2
         minimumFractionDigits = 2
